Move utility into package
