+++ packages/kbn-es-query/index.ts
@@ -116,6 +116,7 @@ export {
   nodeTypes,
   toElasticsearchQuery,
   escapeKuery,
+  escapeQuotes,
 } from './src/kuery';
 
 export {
