+++ packages/kbn-es-query/src/kuery/index.ts
@@ -23,5 +23,5 @@ export const toElasticsearchQuery = (...params: Parameters<typeof astToElasticse
 export { KQLSyntaxError } from './kuery_syntax_error';
 export { nodeTypes, nodeBuilder } from './node_types';
 export { fromKueryExpression, toKqlExpression } from './ast';
-export { escapeKuery } from './utils';
+export { escapeKuery, escapeQuotes } from './utils';
 export type { DslQuery, KueryNode, KueryQueryOptions, KueryParseOptions } from './types';
